Small rewrite aroudn the MoT's response to data request
I found it difficult to read this part the first time, so this rewrite
is inpart to order the sentences so I would have read them more easily
first time. And ot me, the assertion about the MoT's statement makes the
most sense after the statement.

Next, in the sentences above, there are several uses of 'as', so the
additional one "As data people" ended up sounding clunkier than it
should have. So I removed the 'as data people', and several other
changes suggested themselves from that. The result is the sentence in
the commit.